refactor(game): cut over quest views into package modules
